requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.31
Werkzeug==2.3.8
reportlab==4.2.2
pandas==2.2.2
openpyxl==3.1.5